+++ internal/service/organization.go
@@ -3,8 +3,8 @@ package service
 import (
 	"context"
 
-	"github.com/c1emon/lemontree/httpx"
 	"github.com/c1emon/lemontree/model/org"
+	"github.com/c1emon/lemontree/pkg/httpx"
 	"github.com/pkg/errors"
 )
